+++ find_work/cli/bugzilla.py
@@ -4,32 +4,21 @@
 
 """
 CLI subcommands for everything Bugzilla.
-
-This Python module also defines some regular expressions.
-
-``isodate_re`` matches ISO 8601 time/date strings:
-
->>> isodate_re.fullmatch("2024") is None
-True
->>> isodate_re.fullmatch("20090916T09:04:18") is None
-False
 """
 
-import json
-import re
-import time
 import warnings
-from collections.abc import Iterable
+from collections.abc import Collection
+from datetime import datetime
 from typing import Any
-from xmlrpc.client import DateTime
 
 import click
 import gentoopm
+import pydantic_core
 from tabulate import tabulate
 
 from find_work.cache import (
-    read_json_cache,
-    write_json_cache,
+    read_raw_json_cache,
+    write_raw_json_cache,
 )
 from find_work.cli import Message, Options, ProgressDots
 from find_work.constants import BUGZILLA_URL
@@ -45,40 +34,24 @@ with warnings.catch_warnings():
     import bugzilla
     from bugzilla.bug import Bug
 
-isodate_re = re.compile(r"\d{4}\d{2}\d{2}T\d{2}:\d{2}:\d{2}")
 
-
-class BugEncoder(json.JSONEncoder):
-    def default(self, o: Any) -> Any:
-        if isinstance(o, DateTime):
-            return o.value
-        return json.JSONEncoder.default(self, o)
-
-
-def as_datetime(obj: dict) -> dict:
-    result: dict = {}
-    for key, value in obj.items():
-        # FIXME: every matching string will be converted to DateTime
-        if isinstance(value, str) and isodate_re.fullmatch(value):
-            result[key] = DateTime(value)
-            continue
-        result[key] = value
-    return result
-
-
-def _bugs_from_json(data: list[dict]) -> list[Bug]:
+def _bugs_from_raw_json(raw_json: str | bytes) -> list[Bug]:
+    data: list[dict] = pydantic_core.from_json(raw_json)
     with requests_session() as session:
-        bz = bugzilla.Bugzilla(BUGZILLA_URL, requests_session=session)
+        bz = bugzilla.Bugzilla(BUGZILLA_URL, requests_session=session,
+                               force_rest=True)
         return [Bug(bz, dict=bug) for bug in data]
 
 
-def _bugs_to_json(data: Iterable[Bug]) -> list[dict]:
-    return [bug.get_raw_data() for bug in data]
+def _bugs_to_raw_json(data: Collection[Bug]) -> bytes:
+    raw_data = [bug.get_raw_data() for bug in data]
+    return pydantic_core.to_json(raw_data, exclude_none=True)
 
 
 def _fetch_bugs(options: Options, **kwargs: Any) -> list[Bug]:
     with requests_session() as session:
-        bz = bugzilla.Bugzilla(BUGZILLA_URL, requests_session=session)
+        bz = bugzilla.Bugzilla(BUGZILLA_URL, requests_session=session,
+                               force_rest=True)
         query = bz.build_query(
             short_desc=options.bugzilla.short_desc or None,
             product=options.bugzilla.product or None,
@@ -93,7 +66,7 @@ def _fetch_bugs(options: Options, **kwargs: Any) -> list[Bug]:
         return bz.query(query)
 
 
-def _collect_bugs(data: Iterable[Bug], options: Options) -> list[BugView]:
+def _collect_bugs(data: Collection[Bug], options: Options) -> list[BugView]:
     if options.only_installed:
         pm = gentoopm.get_package_manager()
 
@@ -105,7 +78,7 @@ def _collect_bugs(data: Iterable[Bug], options: Options) -> list[BugView]:
             if package not in pm.installed:
                 continue
 
-        date = time.strftime("%F", bug.last_change_time.timetuple())
+        date = datetime.fromisoformat(bug.last_change_time).date().isoformat()
         item = BugView(bug.id, date, bug.assigned_to, bug.summary)
         result.append(item)
     return result
@@ -117,12 +90,11 @@ def _list_bugs(cmd: str, options: Options, **filters: Any) -> None:
 
     options.say(Message.CACHE_LOAD)
     with dots():
-        cached_data = read_json_cache(options.cache_key,
-                                      object_hook=as_datetime)
-    if cached_data is not None:
+        raw_data = read_raw_json_cache(options.cache_key)
+    if raw_data:
         options.say(Message.CACHE_READ)
         with dots():
-            data = _bugs_from_json(cached_data)
+            data = _bugs_from_raw_json(raw_data)
     else:
         options.vecho("Fetching data from Bugzilla API", nl=False, err=True)
         with dots():
@@ -132,8 +104,8 @@ def _list_bugs(cmd: str, options: Options, **filters: Any) -> None:
             return
         options.say(Message.CACHE_WRITE)
         with dots():
-            json_data = _bugs_to_json(data)
-            write_json_cache(json_data, options.cache_key, cls=BugEncoder)
+            raw_json = _bugs_to_raw_json(data)
+            write_raw_json_cache(raw_json, options.cache_key)
 
     bumps = _collect_bugs(data, options)
     if len(bumps) != 0: